sc-6231: move fit_engine_image to base.rs so the backend-candle build resolves it
sc-6139 added `fit_engine_image` (+ `fit_rgb` / `contain_box`) to the macOS-only
image_jobs/flux2.rs, but video_jobs.rs (candle video I2V / First->Last resolve) and
the candle edit handlers call `crate::image_jobs::fit_engine_image` -- so the
`--features backend-candle` lane failed to compile (E0425). The default PR CI is
candle-blind (backend-candle is workflow_dispatch-only), so it merged broken.

Move the three pure image-fit helpers to image_jobs/base.rs, which is included under
any(target_os=macos, all(not(macos), backend-candle)) -- exactly the lanes whose
callers need them -- keeping them in the same `image_jobs` module (include! is
textual). No `#[cfg]` on the moved fns; availability follows base.rs's include cfg.
`should_fit_edit_source` stays in flux2.rs (MLX-only). No macOS behavior change.
Validated: clippy -p sceneworks-worker --features backend-candle -- -D warnings clean.
